Move file and simplify status checks
