+++ extensions/ql-vscode/src/remote-queries/shared/remote-query-result.ts
